feature: sensor: add force sensor
VIOLIN-6223

A sensor of this type measures the force on it, and additionally
compares the force with one or more specified thresholds. The sensor
can output the force value directly. Moreover, it's usually applied
as a press key. In that case, when it detects a force greater than
some given threshold, a corresponding event is reported.

Change-Id: I07840e29e0ba9cb39b776a089a2f191dbcf6b351
Signed-off-by: jinxudong <jinxudong@xiaomi.com>